+++ src/BioBits/insert_size_decision.jl
@@ -62,3 +62,47 @@ end
 
     (r1_error | r2_error) & !not_false_positive
 end
+
+"""
+    one_bp_check(r::LongDNASeq, a::LongDNASeq, nremain::Int64, length_to_check::Int64)
+
+v3.0.0: When finishing matching, Atria might have 1 bp offset because of insert size decision. Check 1 bp offset of reads at adapter (`a`) position (`nremain + 1`) to adapter is necessary.
+
+Return best nremain::Int64.
+"""
+@inline function one_bp_check(r::LongDNASeq, a::LongDNASeq, nremain::Int64, length_to_check::Int64)
+    n = length(r)
+    if nremain >= n - 3 ## no need to check adapter when no adapter.
+        return nremain
+    end
+    nmatch = unsafe_seq_identity(r, a, nremain + 1, 1, length_to_check)
+    nmatch_left = unsafe_seq_identity(r, a, nremain, 1, length_to_check)
+    nmatch_right = unsafe_seq_identity(r, a, nremain + 2, 1, length_to_check)
+    if nmatch >= nmatch_left
+        if nmatch >= nmatch_right
+            nremain
+        else
+            nremain + 1
+        end
+    else
+        if nmatch_left >= nmatch_right
+            nremain - 1
+        else
+            nremain + 1
+        end
+    end
+end
+
+@inline function unsafe_seq_identity(a::LongDNASeq, b::LongDNASeq, ia::Int64, ib::Int64, max_check::Int64; max_a::Int64 = length(a))
+    nmatch = 0
+    ncheck = 0
+    @inbounds while ncheck < max_check && ia <= max_a
+        if a[ia] == b[ib]
+            nmatch += 1
+        end
+        ncheck +=1
+        ia += 1
+        ib += 1
+    end
+    nmatch
+end